Include loader.js first in vendor-static
There are two module registries used by Ember:

- an internal registry with special handling of Ember-specific assets,
  which in pre-Ember 3.27 was used for Ember-internal modules via the
  `Ember.__loader`
- the regular `define`/`require` registry

As of 3.27+, Ember attempts to use the latter and only falls back to the
Ember-specific if it is missing, but this means that if `loader.js` has
not yet loaded (as in the case of the `vendor-static.js` file generated
via this addon), modules can get registered into the separate registries
and therefore not be resolveable at runtime. Pushing `loader.js` itself
into `vendor-static.js`, and making sure that `ember-vendor-split` runs
after `loader.js` does in Ember's build pipeline, works around this by
guaranteeing that it sets up the primary/ambient registry before Ember's
own special sauce happens.

(In the future, after full Embroider adoption, all of this will become
completely unnecessary because there will be much better tools for
defining bundles such that caching will work as we want.)